feat: start top tracks activity
ArtistSearchActivity

In artist search activity, removed the toast and replaced with intent to
start top tracks activity. Passing the artist name as extra.

TopTracksActivity

Built out view components for the top tracks activity:

* list view (top tracks results - dummy data for now)
  * album thumbnail (image view)
  * linear view (to group track & album names)
    * track name (text view)
    * album name (text view)

Wired up the on item select on the list view. Currently just showing
toast with the track and album names.